+++ src/features/Auth/SignInForm/SignInForm.tsx
@@ -6,7 +6,7 @@ import Input from '@/components/atoms/Input/Input';
 import PasswordInput from '@/components/atoms/Input/PasswordInput';
 import Button from '@/components/atoms/Button/Button';
 import FormField from '@/components/molecules/FormField/FormField';
-import styles from './SignInPage.module.scss';
+import styles from './SignInForm.module.scss';
 
 export default function SignInForm() {
   const {
