+++ src/pages/Auth/ChangePasswordPage/ChangePasswordPage.module.scss
