+++ src/pages/Auth/SignInPage/SignInPage.module.scss
@@ -9,10 +9,6 @@
   margin: auto;
 }
 
-.form {
-  width: 100%;
-}
-
 .title {
   font-size: $font-size-xxxlarge;
   font-weight: $font-weight-bold;
@@ -24,26 +20,6 @@
   font-size: $font-size-xsmall;
   font-weight: $font-weight-bold;
   color: $gray-6;
-}
-
-.options {
-  margin-bottom: 0.5rem;
-  font-size: $font-size-xsmall;
-}
-
-.checkbox {
-  display: flex;
-  align-items: center;
-  gap: 0.1rem;
-}
-
-.checkbox input[type='checkbox'] {
-  transform: translateY(-0.05rem);
-}
-
-.loginBtn {
-  width: 100%;
-  margin-bottom: 1rem;
 }
 
 .links {
